add io register definition file with documentation
I painstakenly copied over the snes wiki IO definition pages, made them into
comments, and added asm variables for them. Not just because I am into drone
mind torture, but also because this way it's an easy shortcut to the docs for
the register you're working with by jumping to the definition in your editor. Or
in the case of the VS Code ASM plugin I'm using, it will show a little doc
string popup when hovering over the name. Which currently works so-so.