use Tiled for maps, use custom script for generating map bin
I'd like to be able to be more flexible with creating maps, instead of how we're
currently we're running the whole conversion process through SuperFamiconv
through pngs of the map. I wanted to only use Aseprite, but it turns out that
currently the only way to export tile map information doesn't export x/y tile
flipping, and the API documentation isn't clear enough on how to extract
this. It's also not obvious how to translate maps back into aseprite.

Since Tiled's .tmj files are a sensible and easy to work with format, and Tiled
is of course purpose built to make maps with, this seems like the best option.

So workflow is: make tiles in Aseprite. Export tile sheet as png for
SuperFamiconv to make char, palette and map data. Use tile sheet png for
tiled to make maps, and store as .tmj. Use conversion script taking .tmj, and
taking SuperFamiconv map data to make map bins to include in assembler.

The reason we need SuperFamiconv map data in our script is so we can extract the
color palette data for the tiles, which we'll use in our own map tiles.

Using SuperFamiconv map data is TBD as the simplistic tile setup we have now
only use a couple of colors from the first palette, and thus is always 0.