move sprite: add and use joypad sprite callbacks, use oam mirror
we already had an oam mirror, and update fns for it, but the sprite we showed on
screen was directly written to vram. So instead first set sprite in the mirror,
then on every frame copy all of oam to vram.

the sprite callbacks are a mirror so to speak of the bg callbacks. also write bg
position values to vram at end of fn