actually update vram at beginning of vblank, consistent a/i api
Previously setting the right a/xy byte width was a bit hodge podge. I'm gonna go
with the callee being responsible for the right initial setup and restore. It
has more context so it can judge better, at the detriment of the programmer
juggling more context in their head and a bit more code bloat. Let's see how it
turns out.