initial collision checking: make map, check sprite upper left point
For this initial setup, we make a collision map out of all visible tiles. One
byte per sprite (will hold additional data in the future). We get the data from
the `.tmj` json file.

We collapse x and y of the sprite on the map to a tile coordinate, and look it
up.

Thus far we only check the x and y sprite point. We should check all four
corners.