we're now going from idle to a parabola jump up and down
Not much fancy logic yet. We can only jump up and we come down. The old
collision checking code will stop us from crashing through the ground.

We do this by using sub-pixels of one nibble. I was thinking of just using a
whole byte as it kinda comes for free, but that limits you to x/y axis of 256,
so let's not get box ourselves in like that for now.

But now our velocity can potentially exceed a pixel per frame, we should redo
the collision checking code to use a couple of points, and push us back to just
above the edge of when we crashed through the floor. Or ceiling or sides.

So we should also be aware of what way we're moving, and what points we care
about to check.

When walking on a surface, We should also continuously check if we're falling
off.

Lots to do!

Btw the code should work for 4 players at a time. I should really add 3 more
sprites and see if it all works.