start of using bounding box for collisions
Instead of checking for overlapping tiles and if so, revert to original state,
we want to check arbitrary bounding box (bb) points (6 in this case) within the
sprite withthe bg. We also want to move back a sensible amount of pixels,
depending on how far the bb point sits within the sprite. And this will depend
on what way we're moving.

For now we first calculate how far the top left corner sits within the collision
map, and then we only check the bottom left point for collisions. We still
revert to moving back to previous state to resolve collisions, and we don't
change momentum and state, which we should.