start handling collisions correctly. also clamp to max vert velo
and also set us uglily back into idle when hitting the ground

It's all a bit messy and experimental. Currently only jumping straight up and
down works, and there's an off by one. I defined a couple of other movement
directions for which the code is very wrong. The new collsion text is out of
date.

But it took me some time to get this right, and I just want to save this before
bedtime.