move left and right, with sensible acceleration/deceleration
Every frame we move in a certain direction we add or decrease our
velocity. Depending on what direction you go in and what direction you're
pressing, you'll go either quicker or slower.

- Going in same direction you're pressing, one sub-pixel per frame is added to
velocity.
- Pushing in opposite direction of movement, 5 sub-pixels are subtracted.
- Not pushing in any direction, 2 sub-pixels are subtracted.

Currently we're only moving when we're in the run-state. We should be able to
move when we're either in the run-state or jump state (which are immediately
activated when being in idle state). You should also be able to jump when in
run state. Currently we can only jump when in idle state.