Actually hook up multiple playable player sprites
3 of the 4 can be moved around when using the multitap interface.
The last one needs to be steered with raw input. I might or might not go to the
effort of implementing that. Not sure if 4 is a crowd, seeing all 4 would need
to sprout bullets with their own overhead. But more chaos would also be more
fun. But then again, where does one find so many people to play with?