use a virtual fn pointer for accessing specialized sprite fns
Only one fn for now. Need to do one more to be able to now use the sprite class
for other types of sprites besides players.

The way this is implemented is through a virtual fn table pointer stored in the
sprite struct. Unfortunately actually calling the fn is rather ugly due to the
65816 not really having a better solution for fn calling. Or it seems to me to
be the case:

    ldy #sprite_vtable::collision_end_callback
    lda (sprite::vptr), y ;; so sprite::x is indexing from start of dp
    sta .loword(W0) ;; W0 is a scratch register at $0000
    ldx #0
    jsr (W0, x)