added collision point no collisions callback
In the process, made indirect calling asm sequence a bit more elegant. Shoutout
to oziphanto from the Discord channel for coming up with the general idea.

This should be enough abstracting away to pave the way for adding non-player
sprites that also need to do background collision checks, and non-player sprites
in general.