add concept of bullets and do *something* on collision
So just added basic plumbing for bullets, which at the same time is the first
non-player sprite, so it's a good test to see if the previous abstracting
patches work.

Turns out they worked :) I've added two bullet sprite graphics, added collision
detection callbacks, and currently the bullet just halts when it hits a wall.

There is just one bullet collision point. Which might be a bit little. Expanding
to 4 points in future patch, as it's better to bloat out to fancy processing, as
we can always scale back. Scaling up is usually more problematic.