Implemented bullets bouncing off walls
For player collisions when moving diagonally into a background square, I
shortcutted the logic to always choose snapping vertically, but with ball-like
physics, you can't do that, and you need to properly check if you can actually
move into open space, or if you're blocked by a wall in any direction.

So now we do proper checking of this, which luckily isn't too expensive as we
saved what square we're ending up with for the points we iterate over, so we can
move a square into any direction and check the collision map. It is a bit
hasslesome as you need to do almost the same thing for all four diagonal
directions, and you can't really rely on function calls because of this slight
changing. It's just not worth it. I did need to put the cardinal direction
snapping in their own fns, as potentially more than one needs to be called from
the diagonal direction stuff.

I'm just glad that the general collision logic seems robust enough to handle
the rewrite. And nothing seems to be falling over... yet. Yay!