use arithmetic right shift equivalent when calculating collisions
It is possible for the top-right part of a sprite to fall of the map into
negative territory. This will later be adjusted when we patch this up for the
actual collision points on the sprite.

When doing logical shifts to the right to loose subpixels and change into
collision map coords, we make it a different nr altogether. So do arithmatic
shifts instead.