use signed muliplication when calculating collision y offset
Much like the previous commit, when using unsigned multiplication we get into
trouble when the left upper corner of a sprite is into negative y
territory. We'll try to multiply by that negative amount which is interpreted as
positive, and nothing makes sense anymore.

Luckily we do have signed hardware multiplication as well. This now halves the
amount of playing field that we can check collisions on without doing more work
regarding either adding another layer of coordinate abstraction or doing some
immediate result multiplication things.

For now we're fine like this. Let's see what kind of game we're making before we
expend more effort on this.