do sprite animations for running and remember sprite dir for other states
- Turns out knowing the current way we're pushing the controller isn't enough to
know where we're facing. For example if we're pushing up we still need to know
which direction we were facing before that. Or if/when we implement clinging our
facing direction is always the opposite of the dir we cling to. So we introduce
a stance dir variable.
- We need to do some offset changes from where the bullet comes from to line up
with our nozzle, as before we just shot from the middle point of the sprite.
- Animation speed is very crudely calculated by shifting the anim speed var of
the sprite to the right a couple of bits and masking to the amount of animations
we have (which of course need to be a power of 2).