encode tile select screen element binary data
For the convert-tmj.py script the Tiled tile select png is basically encoded by
the tile-schema doc. convert-tmj.py already creates a tile element
representation that we can just iterate over to generate a SNES tile map from
(admittedly, I thought this one out ahead of time).

So that's what we do here. And besides that we create a bunch of data for the
help screen to know how far we can move the cursor around to hit the bounds, and
we encode the names of the tile rows, that at some point we will use for some
help text, so people aren't too lost when making custom maps.

Essentially what we should end up with is a recreation of our Tiled map making
workflow on the SNES itself, with the benefit instant feedback.