add dma_to_wram fn, and do collisions from wram
This is the first step towards decoupling the current playing field from being
static and in ROM. So we're copying over collision data from rom into wram, and
we update all the collision references to read from wram now.

Once we also do this for tile map data as a staging area, we can start changing
the data in ram, and play around with it. This also opens the door to a changing
environment.