draw tile select screen
We're piggy-bag on the mangingling we already do in convert-tmj.py to create a
map on bg 2 with the same tiles and tile layout that we already use in
Tiled. This map is used for a tile select screen to be able to change the map as
you play.

We're actually recreating this from the data we parsed from the tiled
json, instead of using the png. This allows us to also add borders and be able
to offset it.

Yea so we also have some border creation routines now, which we probably could
reuse somewhere else as well. The code for creating the select screen is a bit
intermixed with the other logic, but reconstructing the rows and columns in
another place seems tedious, and not worth it.

This work on creating the map in python seems to suggest that this might be a
sensible way to create other maps.

The current map is made for BG1, which has the benefit of being able to reuse
the BG2 tiles. Between manging tiled, and doing more bespoke things in code, we
are slowly creating something good here, without having to resort to creating
magic PNGs that we feed to SuperFamiconv. Which I'm not a fan of as you're
losing control, and it will be more bloated and cumbersome to work with, within
the SNES itself.

In general I would expect making menus etc in BG3, but for the tiles this is not
really feasable. I would like the tile select screen be overlaid over the map,
as it feels more immediate and connected than cutting away to a completely
seperate screen. However there's no way to do this with the screen overlay
schema that the SNES uses. BG3 either goes over layers 1 and 2 or it goes below
both.

Currently the see-through BG1 tiles to choose show BG1 below the select
screen. I think the best way to cope with this is to overwrite the BG2 parts
below the screen with uniform black or gray tiles, and then restore it again
when we move away from the menu. This means that we need to save and restore the
previous state of the tiles by maintaining its state in wram, but that was
something that seemed sensible already before this.