fixed some slight bugs in tile graphics + some python script tweaks
the tile graphics had a gray tile drawn at unexpected place, which threw off
tile generation in subtle way (tile was same as tile generated later. if tile
wasnt used it would be discarded, but if later tile was used, it would be
generated instead of later tile).

Also removed redundency in python script and removed tile used for background
from choosable palette. that was still a holdout from a time when we didn't have
a solution for using tile outside of palette shown to end-users.

Plus some cleaning up of python script. Things seem to work sanely now.