Split out `demo.asm` across logical bounds
`demo.asm` felt way too unwieldy for a good while. So we split out the file into
sensible chunks. Perhaps `player.inc` is a bit on the big size, and it might
make sense to do something better with putting the game actions in their own
file in some way. Let's see how workfull things are, and adjust then.