Add a font
I added the boxy bold font by Clint Bellanger, which was subequently modified by
cemkalyoncu, william.thompsonj, and usr_share.

See https://opengameart.org/content/boxy-bold-truetype-font

I also edited the font a bit. It was blown up by a factor of 9, so I scaled it
back, and I changed some of the characters both to fit an 8x8 tile and because
of stylistic disagreements :)

Also the @ wasn't included, so I made a new one for it myself.

Also the tiles were almost, but not completely ASCII ordered, which we need to
easily map from text, so I rearranged all the characters.

For now I used the SuperFamiconv map output to print them all. SuperFamiconv
doesn't do priority bits, so we're stuck with it being behind all the other
graphics, but this is just a test, so that's fine.