gen font map in python to set prio bit and gen font maps from string
You can't instruct SuperFamiconv to set prio bits in the map entries, so gen the
font map in Python.

And while we're add it, add a function to gen a font map from a string at
compile time. Use this to generate the names of the select rows, instead of the
raw string ascii values.