Implement asm knowledge of tile props, use for out of bounds
I'd like for all tiles to be editable, but that means that border tiles can be
turned into empty tiles, and that means that maps can be made that allow players
to go out of bounds.

To combat this I want the sides and the top of the playing field to be
registered as non-go-zone. I've implemented this by now registering the tile
properties when consulting the collision map, and by creating a new kill block
type, that is registered as a prop of the out of bounds area.

Now whenever an entity overlaps with out of bounds, they need to do something
with this info. Currently we frieze the whole game, which is a bit over the top
of course.