We register out of bounds for all 4 sides
For the bottom, we paste 3 more lines to the playing field, so when players are
falling off the map and we do collision checking, the sprite is seen falling off
screen, which I think is a nice touch. This way we can add an extra hazard at
the bottom, which naturally makes sense, and which we can't really do on the
sides or the top.