Set top x pos bit for menu caret, and center initial draw caret pos
Previously the caret position was off by one, as for the first column, the caret
sprite x position should be -1, thus negative, as the caret is one pixel bigger
on all sides of a tile. And I was too lazy to write code to set the top X bit in
the `extra bits` OAM section.

I also centered the initial draw overlay caret position on the screen, to get
you quicker to the place you actually want to be.