bugfix: in coll_left_up handler, move down an actual row of tiles
In the coll_left_up collision handler, we move down a row of tiles and check if
in the row under the collision tile is another collision tile.

Unlike in the other similar handlers, because of a typoe, we only move 16 places
in the collision map, instead of 32.

This patch fixes that.