maintain queue of active bullets
The issue with having multiple bullets in flight for a single player is that
some bullets will last longer than others, and so you'll need to maintain the
set of bullets sensibly and quickly find a bullet that is available. Or know
that none is available at all.

I first tought to use a list for maintaining this, but then figured it's
probably easier to just use an array, with a pivot to mark where where in the
array the line is between the active and inactive bullets.

This code just initializes the array of bullet references (straight in the
player struct) and maintains the queue when handing out a reference which then
becomes active. Turns out that handing out the reference is the cheap part.

Thinking of how to implmement returning used bullets to the queue, to make it
quick(er) (not walking the array to find the offset of the bullet), you'd need
to store its position in the bullet itself, and maintain it when you move about
the pointers, which in the end would be about as costly as maitaining a doubly
linked list. In any case, it's a fun problem to ponder.