Clean up idle bullet logic
Up till now bullets haven't done very much with their move_state field and it
wasn't maintained properly. As there was no way do differentiate between active
and inactive bullets. All were processed all the time. This is fine as this made
me understand the bounds of how much resources we were eating in the profiler.

But this can create some issues as we're also doing collision detection on them,
which will give garbage results, and it can hinder debugging as you always have
to interpret and move past inactive bullets.

So I've closed the loop on this. Bullet move state is correctly updated, and we
don't process bullets for collisions or movement when they're idle.

This wins us 20 percent of processing power when they're idle (in 5 bullet per
player configuration), but we should account for max bullets in flight when
thinking of total processing power budget.