Make player <> bullet collisions more precise
Before we would only check new bullet positions against players. Now we check
start bullet position and middle point of bullet between old and new position.

This because we have both a moving player and a moving bullet, so it's very
possible that if they move towards eachother, they will not overlap and the
bullet will just go straight through. Or when moving diagonally it's easy for
bullet and player to just miss eachother. Just clipping say the left top of the
player sprite. Using the old and new bullet position in the frame is still not
enough. Using the middle position still isn't perfect, but for now it seems
acceptable enough in my empirical tests.