Give all 4 player sprites a different color
.. by assigning them to a different 16 color palette entry. SuperFamiconv does
the right thing when you ask it not to mess with the palette positions. It will
keep all the entries that you make in the same order, and will keep them all.

So in Aseprite you can simply expand the palette pane to 16 colors wide, so you
can see all palettes in rows under eachother. Aseprite also allows for duplicate
colors so you can assign every 1st color of a palette to the transparent color.

It all worked out great, yay.