Make menu backgrounds with border of arbitrary size
The select tile background we made earlier is a bit special in that it needs to
embed the tiles you want to choose in the menu background layer itself.

Normally we don't have this issue and we can just make an empty background of an
arbitrary size. Which then can be overlayed with text and sprites. Since
sometimes you do want to put sprites over the background this means you can't
use layer 3 for this. So we're stuck with basically the same setup as we have
for the tile select screen.

Which is fine. It's the most flexible setup, but it
does mean we do also have to remap the background layer we haven't even
programmed yet for the main game loop. And so I loose my dream of showing the
whole of the last game in the background while also showing the menu
screen. Just a minor setback.