We got an actual game loop!!!
OMG it only cost about 5 months. This patch was all about sowing up some loose
threads.

Now we press a button to start a game, when we shot everyone to bits we get a
win added to our tally, and when we won we get back to a menu, and we do the
whole thing again!