Set selection caret for players
And allocate two more selection caret sprites. One for levels, and another one
because I know we'll need one pretty soon. Perhaps for wins per match for
example. Groups of 4 sprites are good so we can just reserve a whole sprite top
section of two bits for them.

This part of the code feels a bit hacky. Should clean it up a bit at some point
and make sure all the logic for things is in the right place.

Also we've got some artifact shimmering from writing to vram outside of the
blanking period. We need to clean that up too at some time.