read game input config from memory, so we can configure it
I wanted to support both the most sane way and also a more traditional way just
by swapping changing the key combination. unfortunately that isn't quite so easy
as the metroid way of doing this uses L and R as shoot up and down with the
current direction we're facing. Not such an issue but that needs a bit more
tiny bit more work.