Implement bullet burst mode/configurable bullet rate limiting
Single bullets are bit boring. But a string of them might be fun!

Shoot them all while we can.

Also we can configure the time between bursts.

If we just want one bullet as before, we can just change a variable, and we
still get to configure rate limiting for it.

This gives us some levers to play with.