Players that get hit get some invincibility time
So they don't immediately die when they're hit with 10 bullets.

Have at least the idea of a central place for timers. Otherwise they're
scattered all over the place, which might get confusing. But also because for
example this hit counter wouldn't have a logical place. Since the 'get hit' part
of the code isn't always executed, and then where do you put it, while you still
need to make sure that decrementing the counter is being done in a sensible
place.

So we decrement in a fixed place before anything else is run, so we have
something to build upon.