When fight starts, intro text and do nothing for a bit
starting the action right away isn't pretty as some of the players are still
falling and you want the game to breathe a bit.

So we're adding in a game prologue screen that runs the game, but there's no
gamepad input. And we show a 'start fighting' text to tie it all together.

Took away the hackish `fighting-p` variable to indicate that we're in the
fighting screen, as we should generalize that concept a bit more. As the mode
we're in is the fn that is the current game loop driver. Which if it wants can
run the game game loop as a sub loop. This makes things more understandable than
adding another sub mode of the fighting game loop, which I was at first going to
code up.