make sense of VRAM memory, and sensibly divide it
I took some time to understand how to make sense of the memory layout and how
the program reasons about it. Turns out I had some misconceptions which resulted
in broken behaviour (writing to unexpected regions of VRAM), although that
wasn't visibile in the program.

I also needed to understand how much space we acutually have to work with.

Findings:

- VRAM is reasoned about in words. So it reasons about 32KB of words, instead of
64KB of bytes. So when we tell for example the BG registers where a certain part
of memory is, it wants it in word address format.
- Except for how much (H)DMA should move bytes over. That is
universally (between all the different memory regions) set in bytes.

Below is in bytes, so divide by 2 when actually use in programs
- we have two sections of 8KB sprite sheet data
- BG blocks are 2KB aligned
- BG maps are $800B aligned

Placing these all after eachother, with the biggest alignment requirements
first, makes sure we can be sure what we do is sane and flexible.