Disable automatic writing of wram game map to vram outside of game
As we do some tricks to write to WRAM directly when patching in tiles in edit
mode. If we need to start relying on updating in vblank, we need to do a bunch
of extra work to do that sensibly.

One idea is to divert the patching to vram handling as well, which is probably
more elegant. But for now we gotta go fast.