beginning of actual matches instead of just individual games
Currently whenever someone wins, we just transition to the between games screen
and we move on. However, it'd be nice to have a say street fighter-like match
system, where you need to win a number of games to win the match.

This is the beginning of that. You're dropped directly into a new game after
someone wins, if you win a match, you go to the in-between screen.