Add bullet explode tile prop, and make bullet passthrough tiles
The relationship between bullet and player tile props was somewhat fuzzy, as a
wall type served to both not pass bullets and players, but sometimes you want a
wall to pass bullets but not players, or vice-versa..

So we add the explode tile prop and we decouple bullets from players. For
bullets, we can either bounce, explode or pass through when touching a wall. For
players we just keep it to just either pass-through or being a wall.